Writing more integration tests for the "Edu Article" endpoints

+++ backend/app/features/educational_articles/edu_articles_service.py
@@ -1,12 +1,13 @@
-from fastapi import UploadFile
+from fastapi import HTTPException, UploadFile, status
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
-from app.db.db_schema import EduArticle, VolunteerDoctor
+from app.db.db_schema import EduArticle, EduArticleCategory, VolunteerDoctor
 from app.features.educational_articles.edu_article_models import (
     ArticleDetailedResponse,
     ArticleOverviewResponse,
 )
+from app.shared.s3_storage_interface import S3StorageInterface
 
 
 class EduArticlesService:
@@ -14,6 +15,9 @@ class EduArticlesService:
         self.db = db
 
     def get_article_overviews_by_category(self, category: str) -> list[ArticleOverviewResponse]:
+        if not category in [cat.value for cat in list(EduArticleCategory)]:
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST)
+
         article_overviews = (
             self.db.execute(select(EduArticle.id, EduArticle.title).where(EduArticle.category == category))
             .mappings()
@@ -23,9 +27,24 @@ class EduArticlesService:
 
     def get_article_detailed(self, article_id: int) -> ArticleDetailedResponse:
         article = self.db.get(EduArticle, article_id)
+
         if article is None:
-            raise
-        return ArticleDetailedResponse.model_validate(article, from_attributes=True)
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+
+        author: VolunteerDoctor = article.author
+        full_name: str = (
+            f"{author.first_name}{author.middle_name if author.middle_name is not None else ''}{author.last_name}"
+        )
+
+        return ArticleDetailedResponse(
+            id=article.id,
+            author_id=article.author_id,
+            author=full_name,
+            category=article.category,
+            img_key=None,
+            title=article.title,
+            content_markdown=article.content_markdown,
+        )
 
     def create_article(
         self, category: str, title: str, content_markdown: str, img_data: UploadFile, doctor: VolunteerDoctor
@@ -39,10 +58,12 @@ class EduArticlesService:
         )
         self.db.add(article)
         self.db.flush()
-        # article_img_key: str = S3StorageInterface.put_article_img(article.id, img_data)
-        # article.img_key = article_img_key
+        article_img_key: str = S3StorageInterface.put_article_img(article.id, img_data)
+        article.img_key = article_img_key
         return article
 
     def delete_article(self, article_id: int) -> None:
         article = self.db.get(EduArticle, article_id)
+        if article is None:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
         self.db.delete(article)
